Dependencies: Add support for Python 3.11
Officially add support for Python 3.11 by adding it in the trove
classifiers and run the CI workflow against it.

The code required a single fix to add compatiblity. The `ZipFile` class
in the standard library got the `metadata_encoding` argument added, see:
https://github.com/python/cpython/commit/a25a985535ccbb7df8caddc0017550ff4eae5855
This needed to be added to the `ZipExtra` class which essentially clones
the class of the standard library. Without it, an `AttributeError` would
be thrown when `_RealGetContents` would call `self.metadata_encoding`.